Move footerHeight global SharedValue update to SceneWrapper
This is the right place where we can subscribe to focus events to make
sure the shared value is always correctly the footerHeight for the
currently focused scene. This is because the SceneWrapperFooter is
render in the MenuTabs sometimes, and that means it wont get the focus
events during scene transitions.
