+++ src/components/themed/SceneFooterWrapper.tsx
@@ -1,4 +1,3 @@
-import { useIsFocused } from '@react-navigation/native'
 import React, { useEffect } from 'react'
 import Animated, { interpolate, SharedValue, useAnimatedStyle } from 'react-native-reanimated'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
@@ -23,7 +22,6 @@ export const SceneFooterWrapper = (props: SceneFooterProps) => {
   const { children, noBackgroundBlur = false, sceneWrapperInfo, onLayoutHeight } = props
   const { hasTabs = true, isKeyboardOpen = false } = sceneWrapperInfo ?? {}
   const footerOpenRatio = useSceneFooterState(state => state.footerOpenRatio)
-  const footerHeightShared = useSceneFooterState(state => state.footerHeight)
 
   const safeAreaInsets = useSafeAreaInsets()
   const maybeInsetBottom = !hasTabs ? safeAreaInsets.bottom : 0
@@ -44,18 +42,6 @@ export const SceneFooterWrapper = (props: SceneFooterProps) => {
     const footerHeight = layout.height - maybeInsetBottom
     onLayoutHeight(footerHeight)
   }, [layout, maybeInsetBottom, onLayoutHeight])
-
-  // Set the global shared value for the footerHeight so that way the
-  // background in the MenuTabs can translate accordingly
-  const isFocused = useIsFocused()
-  useEffect(() => {
-    if (isFocused) {
-      footerHeightShared.value = layout?.height ?? 0
-      return () => {
-        footerHeightShared.value = 0
-      }
-    }
-  }, [footerHeightShared, isFocused, layout?.height])
 
   //
   // Render
